+++ internal/model/user.go
@@ -1,0 +1,6 @@
+package model
+
+type User struct {
+	UID  string `json:"uid"`
+	Name string `json:"name"`
+}